backend/pyproject.toml
[tool.poetry]
name = "qual"
version = "0.1.0"
description = "这是一段描述"
authors = ["kaluluosi111 <kaluluosi@gmail.com>"]
readme = "ABOUT.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
pydantic-settings = "^2.0.3"
coloredlogs = "^15.0.1"
bcrypt = "^4.0.1"
pip = "^23.2.1"
install = "^1.3.5"
itsdangerous = "^2.1.2"
httpx = "^0.25.0"
passlib = "^1.7.4"
fastapi-sso = "^0.7.2"

[tool.black]
line-length = 88

[tool.ruff]
ignore = ['E501']

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
ruff = "^0.0.289"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
